move cache related structure and functions to authenticate_cache file
